Fix server side use of window
